ALM defects

+++ src/features/Container/Protocols/CustomComponents/PDTable/index.js
@@ -6,7 +6,6 @@ import ButtonGroup from 'apollo-react/components/ButtonGroup';
 import Tooltip from 'apollo-react/components/Tooltip';
 import IconButton from 'apollo-react/components/IconButton';
 import Plus from 'apollo-react-icons/Plus';
-import EmptyColumnCells from './Components/EmptyColumns';
 import DisplayTable from './Components/Table';
 import { tableOperations } from './Components/dropdownData';
 import { addColumn, addRow, deleteColumn, deleteRow } from './utils';
@@ -71,7 +70,7 @@ const formattableData = (data) => {
 
 const confirmText = 'Please confirm if you want to continue with deletion';
 
-function PDTable({ data, segment, activeLineID, lineID }) {
+function PDTable({ data, segment, activeLineID, lineID, setIsTableChanged }) {
   const [updatedData, setUpdatedData] = useState([]);
   const [footNoteData, setFootnoteData] = useState([]);
   const [columnLength, setColumnLength] = useState();
@@ -103,7 +102,9 @@ function PDTable({ data, segment, activeLineID, lineID }) {
     const cloneData = [...updatedData];
     cloneData[rowIndex][columnIndex].content = content;
     setUpdatedData(cloneData);
+    setIsTableChanged(true);
   };
+
   const handleColumnOperation = (operation, index) => {
     if (operation === tableOperations.addColumnLeft) {
       const newData = addColumn(updatedData, index);
@@ -115,11 +116,16 @@ function PDTable({ data, segment, activeLineID, lineID }) {
       setUpdatedData(newData);
       setColumnLength(Object.keys(newData[0]).length);
     } else if (operation === tableOperations.deleteColumn) {
-      const newData = deleteColumn(updatedData, index);
-      setUpdatedData(newData);
-      setColumnLength(Object.keys(newData[0]).length);
+      // eslint-disable-next-line
+      if (confirm(confirmText)) {
+        const newData = deleteColumn(updatedData, index);
+        setUpdatedData(newData);
+        setColumnLength(Object.keys(newData[0]).length);
+      }
     }
+    setIsTableChanged(true);
   };
+
   const handleRowOperation = (operation, index) => {
     if (operation === tableOperations.addRowAbove) {
       const newData = addRow(updatedData, index);
@@ -134,6 +140,7 @@ function PDTable({ data, segment, activeLineID, lineID }) {
       setIsModal(true);
       setSelectedIndex(index);
     }
+    setIsTableChanged(true);
   };
 
   const addFootNote = () => {
@@ -145,6 +152,7 @@ function PDTable({ data, segment, activeLineID, lineID }) {
       },
     ]);
   };
+
   const handleSave = () => {
     const content = {
       ...segment.content,
@@ -152,6 +160,7 @@ function PDTable({ data, segment, activeLineID, lineID }) {
       AttachmentListProperties: footNoteData,
     };
     setTableSaved(true);
+    setIsTableChanged(false);
     dispatchSectionEvent('CONTENT_UPDATE', {
       type: CONTENT_TYPE.TABLE,
       lineID,
@@ -159,6 +168,7 @@ function PDTable({ data, segment, activeLineID, lineID }) {
       content,
     });
   };
+
   const isEditable = () => {
     return lineID === activeLineID && !tableSaved;
   };
@@ -183,10 +193,12 @@ function PDTable({ data, segment, activeLineID, lineID }) {
               },
               {
                 label: 'Delete',
-                onClick: () =>
+                onClick: () => {
                   dispatchSectionEvent('CONTENT_DELETED', {
                     currentLineId: activeLineID,
-                  }),
+                  });
+                  setIsTableChanged(false);
+                },
               },
             ]}
           />
@@ -216,19 +228,7 @@ function PDTable({ data, segment, activeLineID, lineID }) {
           />
         </div>
       )}
-      {/* eslint-disable-next-line */}
-      <div
-        className="pd-table-container"
-        ref={tableRef}
-        onClick={() => setTableSaved(false)}
-      >
-        {isEditable() && (
-          <EmptyColumnCells
-            columnLength={columnLength}
-            handleOperation={handleColumnOperation}
-            colWidth={colWidth}
-          />
-        )}
+      <div className="pd-table-container" ref={tableRef}>
         <DisplayTable
           data={updatedData}
           onChange={handleChange}
@@ -237,6 +237,8 @@ function PDTable({ data, segment, activeLineID, lineID }) {
           colWidth={colWidth}
           footNoteData={footNoteData}
           setFootnoteData={setFootnoteData}
+          handleColumnOperation={handleColumnOperation}
+          columnLength={columnLength}
         />
       </div>
       <Modal
@@ -269,4 +271,5 @@ PDTable.propTypes = {
   segment: PropTypes.isRequired,
   activeLineID: PropTypes.isRequired,
   lineID: PropTypes.isRequired,
+  setIsTableChanged: PropTypes.isRequired,
 };
